Added a mechanism to specify a common library path within the configuration.
Prior to this change you could either put the 3rd party
library into the system 'libraries' directory for the arduino software
(typically at /usr/local/share/arduino/libraries) or copy/symlink it
into the [project]/lib directory.

With this change there is a third option for including 3rd party
libraries.  The build command now allows for another library
directory to be provided.  This library directory will be used just
like the prior two.

The option is "--common-library" and it can be used with the build
command or placed within the standard configuration files.